age_classification/requirements.txt
streamlit==1.52.1
opencv-python==4.12.0
numpy==2.2.6
Pillow==12.0.0
scikit-learn==1.3.1
scikit-image==0.21.0
joblib==1.3.2
ultralytics==8.1.1
torch==2.2.0
